add subtasks helper
delete checking helper, put checking functions in task helper
